demo/phonic_chat/requirements.txt
fastrtc
python-dotenv
phonic-python
twilio
